@@ -1,13 +1,13 @@
 {
   "private": true,
   "devDependencies": {
-    "fuse-box": "4.0.0-next.165",
+    "fuse-box": "4.0.0-next.435",
     "parcel": "1.12.4",
-    "rollup": "1.31.1",
-    "rollup-plugin-terser": "5.2.0",
-    "ts-loader": "6.2.2",
-    "typescript": "3.7.5",
-    "webpack": "4.41.6",
+    "rollup": "2.10.9",
+    "rollup-plugin-terser": "6.1.0",
+    "ts-loader": "7.0.5",
+    "typescript": "3.9.3",
+    "webpack": "4.43.0",
     "webpack-cli": "3.3.11"
   }
 }
