@@ -17,6 +17,7 @@ import (
 	"github.com/evanw/esbuild/internal/lexer"
 	"github.com/evanw/esbuild/internal/logging"
 	"github.com/evanw/esbuild/internal/parser"
+	"github.com/evanw/esbuild/internal/printer"
 	"github.com/evanw/esbuild/internal/resolver"
 )
 
@@ -33,7 +34,7 @@ Options:
   --target=...          Language target (default esnext)
   --platform=...        Platform target (browser or node, default browser)
   --external:M          Exclude module M from the bundle
-  --format=...          Output format (iife or cjs)
+  --format=...          Output format (iife, cjs, esm)
   --color=...           Force use of color terminal escapes (true or false)
 
   --minify              Sets all --minify-* flags
@@ -53,6 +54,7 @@ Advanced options:
   --sourcemap=external  Do not link to the source map with a comment
   --sourcefile=...      Set the source file for the source map (for stdin)
   --error-limit=...     Maximum error count or 0 to disable (default 10)
+  --log-level=...       Disable logging (info, warning, error)
 
   --trace=...           Write a CPU trace to this file
   --cpuprofile=...      Write a CPU profile to this file
@@ -75,11 +77,18 @@ Examples:
 type argsObject struct {
 	traceFile      string
 	cpuprofileFile string
+	rawDefines     map[string]parser.DefineFunc
 	parseOptions   parser.ParseOptions
 	bundleOptions  bundler.BundleOptions
 	resolveOptions resolver.ResolveOptions
 	logOptions     logging.StderrOptions
 	entryPaths     []string
+}
+
+func (args argsObject) logInfo(text string) {
+	if args.logOptions.LogLevel <= logging.LevelInfo {
+		fmt.Fprintf(os.Stderr, "%s\n", text)
+	}
 }
 
 func exitWithError(text string) {
@@ -105,10 +114,15 @@ func (args *argsObject) parseDefine(key string, value string) bool {
 		}
 	}
 
+	// Lazily create the defines map
+	if args.rawDefines == nil {
+		args.rawDefines = make(map[string]parser.DefineFunc)
+	}
+
 	// Allow substituting for an identifier
 	if lexer.IsIdentifier(value) {
 		if _, ok := lexer.Keywords()[value]; !ok {
-			args.parseOptions.Defines[key] = func(findSymbol parser.FindSymbol) ast.E {
+			args.rawDefines[key] = func(findSymbol parser.FindSymbol) ast.E {
 				return &ast.EIdentifier{findSymbol(value)}
 			}
 			return true
@@ -139,7 +153,7 @@ func (args *argsObject) parseDefine(key string, value string) bool {
 		return false
 	}
 
-	args.parseOptions.Defines[key] = fn
+	args.rawDefines[key] = fn
 	return true
 }
 
@@ -182,9 +196,6 @@ func (args *argsObject) parseMemberExpression(text string) ([]string, bool) {
 
 func parseArgs(fs fs.FS, rawArgs []string) (argsObject, error) {
 	args := argsObject{
-		parseOptions: parser.ParseOptions{
-			Defines: make(map[string]parser.DefineFunc),
-		},
 		bundleOptions: bundler.BundleOptions{
 			ExtensionToLoader: bundler.DefaultExtensionToLoaderMap(),
 		},
@@ -335,11 +346,13 @@ func parseArgs(fs fs.FS, rawArgs []string) (argsObject, error) {
 		case strings.HasPrefix(arg, "--format="):
 			switch arg[len("--format="):] {
 			case "iife":
-				args.bundleOptions.OutputFormat = bundler.FormatIIFE
+				args.bundleOptions.OutputFormat = printer.FormatIIFE
 			case "cjs":
-				args.bundleOptions.OutputFormat = bundler.FormatCommonJS
+				args.bundleOptions.OutputFormat = printer.FormatCommonJS
+			case "esm":
+				args.bundleOptions.OutputFormat = printer.FormatESModule
 			default:
-				return argsObject{}, fmt.Errorf("Valid formats: iife, cjs")
+				return argsObject{}, fmt.Errorf("Valid formats: iife, cjs, esm")
 			}
 
 		case strings.HasPrefix(arg, "--color="):
@@ -373,6 +386,18 @@ func parseArgs(fs fs.FS, rawArgs []string) (argsObject, error) {
 				return argsObject{}, fmt.Errorf("Invalid JSX fragment: %s", arg)
 			}
 
+		case strings.HasPrefix(arg, "--log-level="):
+			switch arg[len("--log-level="):] {
+			case "info":
+				args.logOptions.LogLevel = logging.LevelInfo
+			case "warning":
+				args.logOptions.LogLevel = logging.LevelWarning
+			case "error":
+				args.logOptions.LogLevel = logging.LevelError
+			default:
+				return argsObject{}, fmt.Errorf("Invalid log level: %s", arg)
+			}
+
 		case strings.HasPrefix(arg, "--trace="):
 			args.traceFile = arg[len("--trace="):]
 
@@ -404,13 +429,24 @@ func parseArgs(fs fs.FS, rawArgs []string) (argsObject, error) {
 		args.bundleOptions.AbsOutputDir = fs.Dir(args.bundleOptions.AbsOutputFile)
 	}
 
-	if args.bundleOptions.OutputFormat == bundler.FormatNone {
+	// Disallow bundle-only options when not bundling
+	if !args.bundleOptions.IsBundling {
+		if args.bundleOptions.OutputFormat != printer.FormatPreserve {
+			return argsObject{}, fmt.Errorf("Cannot use --format without --bundle")
+		}
+
+		if len(args.resolveOptions.ExternalModules) > 0 {
+			return argsObject{}, fmt.Errorf("Cannot use --external without --bundle")
+		}
+	}
+
+	if args.bundleOptions.IsBundling && args.bundleOptions.OutputFormat == printer.FormatPreserve {
 		// If the format isn't specified, set the default format using the platform
 		switch args.resolveOptions.Platform {
 		case resolver.PlatformBrowser:
-			args.bundleOptions.OutputFormat = bundler.FormatIIFE
+			args.bundleOptions.OutputFormat = printer.FormatIIFE
 		case resolver.PlatformNode:
-			args.bundleOptions.OutputFormat = bundler.FormatCommonJS
+			args.bundleOptions.OutputFormat = printer.FormatCommonJS
 		}
 	}
 
@@ -423,9 +459,6 @@ func parseArgs(fs fs.FS, rawArgs []string) (argsObject, error) {
 		// Write to stdout by default if there's only one input file
 		if len(args.entryPaths) == 1 && args.bundleOptions.AbsOutputFile == "" && args.bundleOptions.AbsOutputDir == "" {
 			args.bundleOptions.WriteToStdout = true
-			if args.bundleOptions.SourceMap != bundler.SourceMapNone {
-				args.bundleOptions.SourceMap = bundler.SourceMapInline
-			}
 		}
 	} else if !logging.GetTerminalInfo(os.Stdin).IsTTY {
 		// If called with no input files and we're not a TTY, read from stdin instead
@@ -442,12 +475,23 @@ func parseArgs(fs fs.FS, rawArgs []string) (argsObject, error) {
 				return argsObject{}, fmt.Errorf("Cannot use --outdir when reading from stdin")
 			}
 			args.bundleOptions.WriteToStdout = true
-			if args.bundleOptions.SourceMap != bundler.SourceMapNone {
-				args.bundleOptions.SourceMap = bundler.SourceMapInline
-			}
 		}
 	}
 
+	// Change the default value for some settings if we're writing to stdout
+	if args.bundleOptions.WriteToStdout {
+		if args.bundleOptions.SourceMap != bundler.SourceMapNone {
+			args.bundleOptions.SourceMap = bundler.SourceMapInline
+		}
+		if args.logOptions.LogLevel == logging.LevelNone {
+			args.logOptions.LogLevel = logging.LevelWarning
+		}
+	}
+
+	// Processing defines is expensive. Process them once here so the same object
+	// can be shared between all parsers we create using these arguments.
+	processedDefines := parser.ProcessDefines(args.rawDefines)
+	args.parseOptions.Defines = &processedDefines
 	return args, nil
 }
 
@@ -501,7 +545,7 @@ func main() {
 			}
 			defer func() {
 				f.Close()
-				fmt.Fprintf(os.Stderr, "Wrote to %s\n", args.traceFile)
+				args.logInfo(fmt.Sprintf("Wrote to %s", args.traceFile))
 			}()
 			trace.Start(f)
 			defer trace.Stop()
@@ -518,7 +562,7 @@ func main() {
 			}
 			defer func() {
 				f.Close()
-				fmt.Fprintf(os.Stderr, "Wrote to %s\n", args.cpuprofileFile)
+				args.logInfo(fmt.Sprintf("Wrote to %s", args.cpuprofileFile))
 			}()
 			pprof.StartCPUProfile(f)
 			defer pprof.StopCPUProfile()
@@ -529,7 +573,7 @@ func main() {
 			// return useful information for esbuild, since it's so fast. Let's keep
 			// running for 30 seconds straight, which should give us 3,000 samples.
 			seconds := 30.0
-			fmt.Fprintf(os.Stderr, "Running for %g seconds straight due to --cpuprofile...\n", seconds)
+			args.logInfo(fmt.Sprintf("Running for %g seconds straight due to --cpuprofile...", seconds))
 			for time.Since(start).Seconds() < seconds {
 				run(fs, args)
 			}
@@ -544,9 +588,7 @@ func main() {
 		}
 	}()
 
-	if !args.bundleOptions.WriteToStdout {
-		fmt.Fprintf(os.Stderr, "Done in %dms\n", time.Since(start).Nanoseconds()/1000000)
-	}
+	args.logInfo(fmt.Sprintf("Done in %dms", time.Since(start).Nanoseconds()/1000000))
 }
 
 func run(fs fs.FS, args argsObject) {
@@ -593,9 +635,7 @@ func run(fs fs.FS, args argsObject) {
 		if err != nil {
 			exitWithError(fmt.Sprintf("Failed to write to %s (%s)", path, err.Error()))
 		}
-		if !args.bundleOptions.WriteToStdout {
-			fmt.Fprintf(os.Stderr, "Wrote to %s (%s)\n", path, toSize(len(item.JsContents)))
-		}
+		args.logInfo(fmt.Sprintf("Wrote to %s (%s)", path, toSize(len(item.JsContents))))
 
 		// Also write the source map
 		if item.SourceMapAbsPath != "" {
@@ -604,9 +644,7 @@ func run(fs fs.FS, args argsObject) {
 			if err != nil {
 				exitWithError(fmt.Sprintf("Failed to write to %s: (%s)", path, err.Error()))
 			}
-			if !args.bundleOptions.WriteToStdout {
-				fmt.Fprintf(os.Stderr, "Wrote to %s (%s)\n", path, toSize(len(item.SourceMapContents)))
-			}
+			args.logInfo(fmt.Sprintf("Wrote to %s (%s)", path, toSize(len(item.SourceMapContents))))
 		}
 	}
 }
