@@ -4,9 +4,10 @@ import (
 	"testing"
 
 	"github.com/evanw/esbuild/internal/parser"
+	"github.com/evanw/esbuild/internal/printer"
 )
 
-func TestImportStarES6Unused(t *testing.T) {
+func TestImportStarUnused(t *testing.T) {
 	expectBundled(t, bundled{
 		files: map[string]string{
 			"/entry.js": `
@@ -27,22 +28,17 @@ func TestImportStarES6Unused(t *testing.T) {
 			AbsOutputFile: "/out.js",
 		},
 		expected: map[string]string{
-			"/out.js": `bootstrap({
-  0() {
-    // /foo.js
-    const foo2 = 123;
+			"/out.js": `// /foo.js
 
-    // /entry.js
-    let foo = 234;
-    console.log(foo);
-  }
-}, 0);
+// /entry.js
+let foo = 234;
+console.log(foo);
 `,
 		},
 	})
 }
 
-func TestImportStarES6Capture(t *testing.T) {
+func TestImportStarCapture(t *testing.T) {
 	expectBundled(t, bundled{
 		files: map[string]string{
 			"/entry.js": `
@@ -63,26 +59,22 @@ func TestImportStarES6Capture(t *testing.T) {
 			AbsOutputFile: "/out.js",
 		},
 		expected: map[string]string{
-			"/out.js": `bootstrap({
-  0() {
-    // /foo.js
-    var exports = {};
-    __export(exports, {
-      foo: () => foo2
-    });
-    const foo2 = 123;
+			"/out.js": `// /foo.js
+const foo_exports = {};
+__export(foo_exports, {
+  foo: () => foo2
+});
+const foo2 = 123;
 
-    // /entry.js
-    let foo = 234;
-    console.log(exports, foo2, foo);
-  }
-}, 0);
+// /entry.js
+let foo = 234;
+console.log(foo_exports, foo2, foo);
 `,
 		},
 	})
 }
 
-func TestImportStarES6NoCapture(t *testing.T) {
+func TestImportStarNoCapture(t *testing.T) {
 	expectBundled(t, bundled{
 		files: map[string]string{
 			"/entry.js": `
@@ -103,22 +95,18 @@ func TestImportStarES6NoCapture(t *testing.T) {
 			AbsOutputFile: "/out.js",
 		},
 		expected: map[string]string{
-			"/out.js": `bootstrap({
-  0() {
-    // /foo.js
-    const foo2 = 123;
+			"/out.js": `// /foo.js
+const foo2 = 123;
 
-    // /entry.js
-    let foo = 234;
-    console.log(foo2, foo2, foo);
-  }
-}, 0);
+// /entry.js
+let foo = 234;
+console.log(foo2, foo2, foo);
 `,
 		},
 	})
 }
 
-func TestImportStarES6ExportImportStarUnused(t *testing.T) {
+func TestImportStarExportImportStarUnused(t *testing.T) {
 	expectBundled(t, bundled{
 		files: map[string]string{
 			"/entry.js": `
@@ -143,28 +131,19 @@ func TestImportStarES6ExportImportStarUnused(t *testing.T) {
 			AbsOutputFile: "/out.js",
 		},
 		expected: map[string]string{
-			"/out.js": `bootstrap({
-  1() {
-    // /foo.js
-    var exports = {};
-    __export(exports, {
-      foo: () => foo2
-    });
-    const foo2 = 123;
+			"/out.js": `// /foo.js
 
-    // /bar.js
+// /bar.js
 
-    // /entry.js
-    let foo = 234;
-    console.log(foo);
-  }
-}, 1);
+// /entry.js
+let foo = 234;
+console.log(foo);
 `,
 		},
 	})
 }
 
-func TestImportStarES6ExportImportStarNoCapture(t *testing.T) {
+func TestImportStarExportImportStarNoCapture(t *testing.T) {
 	expectBundled(t, bundled{
 		files: map[string]string{
 			"/entry.js": `
@@ -189,28 +168,24 @@ func TestImportStarES6ExportImportStarNoCapture(t *testing.T) {
 			AbsOutputFile: "/out.js",
 		},
 		expected: map[string]string{
-			"/out.js": `bootstrap({
-  1() {
-    // /foo.js
-    var exports = {};
-    __export(exports, {
-      foo: () => foo2
-    });
-    const foo2 = 123;
+			"/out.js": `// /foo.js
+const foo_exports = {};
+__export(foo_exports, {
+  foo: () => foo2
+});
+const foo2 = 123;
 
-    // /bar.js
+// /bar.js
 
-    // /entry.js
-    let foo = 234;
-    console.log(exports.foo, exports.foo, foo);
-  }
-}, 1);
+// /entry.js
+let foo = 234;
+console.log(foo_exports.foo, foo_exports.foo, foo);
 `,
 		},
 	})
 }
 
-func TestImportStarES6ExportImportStarCapture(t *testing.T) {
+func TestImportStarExportImportStarCapture(t *testing.T) {
 	expectBundled(t, bundled{
 		files: map[string]string{
 			"/entry.js": `
@@ -235,28 +210,24 @@ func TestImportStarES6ExportImportStarCapture(t *testing.T) {
 			AbsOutputFile: "/out.js",
 		},
 		expected: map[string]string{
-			"/out.js": `bootstrap({
-  1() {
-    // /foo.js
-    var exports = {};
-    __export(exports, {
-      foo: () => foo2
-    });
-    const foo2 = 123;
+			"/out.js": `// /foo.js
+const foo_exports = {};
+__export(foo_exports, {
+  foo: () => foo2
+});
+const foo2 = 123;
 
-    // /bar.js
+// /bar.js
 
-    // /entry.js
-    let foo = 234;
-    console.log(exports, exports.foo, foo);
-  }
-}, 1);
+// /entry.js
+let foo = 234;
+console.log(foo_exports, foo_exports.foo, foo);
 `,
 		},
 	})
 }
 
-func TestImportStarES6ExportStarAsUnused(t *testing.T) {
+func TestImportStarExportStarAsUnused(t *testing.T) {
 	expectBundled(t, bundled{
 		files: map[string]string{
 			"/entry.js": `
@@ -280,28 +251,19 @@ func TestImportStarES6ExportStarAsUnused(t *testing.T) {
 			AbsOutputFile: "/out.js",
 		},
 		expected: map[string]string{
-			"/out.js": `bootstrap({
-  1() {
-    // /foo.js
-    var exports = {};
-    __export(exports, {
-      foo: () => foo2
-    });
-    const foo2 = 123;
+			"/out.js": `// /foo.js
 
-    // /bar.js
+// /bar.js
 
-    // /entry.js
-    let foo = 234;
-    console.log(foo);
-  }
-}, 1);
+// /entry.js
+let foo = 234;
+console.log(foo);
 `,
 		},
 	})
 }
 
-func TestImportStarES6ExportStarAsNoCapture(t *testing.T) {
+func TestImportStarExportStarAsNoCapture(t *testing.T) {
 	expectBundled(t, bundled{
 		files: map[string]string{
 			"/entry.js": `
@@ -325,28 +287,24 @@ func TestImportStarES6ExportStarAsNoCapture(t *testing.T) {
 			AbsOutputFile: "/out.js",
 		},
 		expected: map[string]string{
-			"/out.js": `bootstrap({
-  1() {
-    // /foo.js
-    var exports = {};
-    __export(exports, {
-      foo: () => foo2
-    });
-    const foo2 = 123;
+			"/out.js": `// /foo.js
+const foo_exports = {};
+__export(foo_exports, {
+  foo: () => foo2
+});
+const foo2 = 123;
 
-    // /bar.js
+// /bar.js
 
-    // /entry.js
-    let foo = 234;
-    console.log(exports.foo, exports.foo, foo);
-  }
-}, 1);
+// /entry.js
+let foo = 234;
+console.log(foo_exports.foo, foo_exports.foo, foo);
 `,
 		},
 	})
 }
 
-func TestImportStarES6ExportStarAsCapture(t *testing.T) {
+func TestImportStarExportStarAsCapture(t *testing.T) {
 	expectBundled(t, bundled{
 		files: map[string]string{
 			"/entry.js": `
@@ -370,28 +328,24 @@ func TestImportStarES6ExportStarAsCapture(t *testing.T) {
 			AbsOutputFile: "/out.js",
 		},
 		expected: map[string]string{
-			"/out.js": `bootstrap({
-  1() {
-    // /foo.js
-    var exports = {};
-    __export(exports, {
-      foo: () => foo2
-    });
-    const foo2 = 123;
+			"/out.js": `// /foo.js
+const foo_exports = {};
+__export(foo_exports, {
+  foo: () => foo2
+});
+const foo2 = 123;
 
-    // /bar.js
+// /bar.js
 
-    // /entry.js
-    let foo = 234;
-    console.log(exports, exports.foo, foo);
-  }
-}, 1);
+// /entry.js
+let foo = 234;
+console.log(foo_exports, foo_exports.foo, foo);
 `,
 		},
 	})
 }
 
-func TestImportStarES6ExportStarUnused(t *testing.T) {
+func TestImportStarExportStarUnused(t *testing.T) {
 	expectBundled(t, bundled{
 		files: map[string]string{
 			"/entry.js": `
@@ -415,24 +369,19 @@ func TestImportStarES6ExportStarUnused(t *testing.T) {
 			AbsOutputFile: "/out.js",
 		},
 		expected: map[string]string{
-			"/out.js": `bootstrap({
-  1() {
-    // /foo.js
-    const foo2 = 123;
+			"/out.js": `// /foo.js
 
-    // /bar.js
+// /bar.js
 
-    // /entry.js
-    let foo = 234;
-    console.log(foo);
-  }
-}, 1);
+// /entry.js
+let foo = 234;
+console.log(foo);
 `,
 		},
 	})
 }
 
-func TestImportStarES6ExportStarNoCapture(t *testing.T) {
+func TestImportStarExportStarNoCapture(t *testing.T) {
 	expectBundled(t, bundled{
 		files: map[string]string{
 			"/entry.js": `
@@ -456,24 +405,20 @@ func TestImportStarES6ExportStarNoCapture(t *testing.T) {
 			AbsOutputFile: "/out.js",
 		},
 		expected: map[string]string{
-			"/out.js": `bootstrap({
-  1() {
-    // /foo.js
-    const foo2 = 123;
+			"/out.js": `// /foo.js
+const foo2 = 123;
 
-    // /bar.js
+// /bar.js
 
-    // /entry.js
-    let foo = 234;
-    console.log(foo2, foo2, foo);
-  }
-}, 1);
+// /entry.js
+let foo = 234;
+console.log(foo2, foo2, foo);
 `,
 		},
 	})
 }
 
-func TestImportStarES6ExportStarCapture(t *testing.T) {
+func TestImportStarExportStarCapture(t *testing.T) {
 	expectBundled(t, bundled{
 		files: map[string]string{
 			"/entry.js": `
@@ -497,22 +442,18 @@ func TestImportStarES6ExportStarCapture(t *testing.T) {
 			AbsOutputFile: "/out.js",
 		},
 		expected: map[string]string{
-			"/out.js": `bootstrap({
-  1() {
-    // /foo.js
-    const foo2 = 123;
+			"/out.js": `// /foo.js
+const foo2 = 123;
 
-    // /bar.js
-    var bar = {};
-    __export(bar, {
-      foo: () => foo2
-    });
+// /bar.js
+const bar_exports = {};
+__export(bar_exports, {
+  foo: () => foo2
+});
 
-    // /entry.js
-    let foo = 234;
-    console.log(bar, foo2, foo);
-  }
-}, 1);
+// /entry.js
+let foo = 234;
+console.log(bar_exports, foo2, foo);
 `,
 		},
 	})
@@ -539,19 +480,15 @@ func TestImportStarCommonJSUnused(t *testing.T) {
 			AbsOutputFile: "/out.js",
 		},
 		expected: map[string]string{
-			"/out.js": `bootstrap({
-  1(exports) {
-    // /foo.js
-    exports.foo = 123;
-  },
+			"/out.js": `// /foo.js
+var require_foo = __commonJS((exports) => {
+  exports.foo = 123;
+});
 
-  0() {
-    // /entry.js
-    const ns = __import(1 /* ./foo */);
-    let foo = 234;
-    console.log(foo);
-  }
-}, 0);
+// /entry.js
+const ns = __toModule(require_foo());
+let foo = 234;
+console.log(foo);
 `,
 		},
 	})
@@ -578,19 +515,15 @@ func TestImportStarCommonJSCapture(t *testing.T) {
 			AbsOutputFile: "/out.js",
 		},
 		expected: map[string]string{
-			"/out.js": `bootstrap({
-  1(exports) {
-    // /foo.js
-    exports.foo = 123;
-  },
+			"/out.js": `// /foo.js
+var require_foo = __commonJS((exports) => {
+  exports.foo = 123;
+});
 
-  0() {
-    // /entry.js
-    const ns = __import(1 /* ./foo */);
-    let foo = 234;
-    console.log(ns, ns.foo, foo);
-  }
-}, 0);
+// /entry.js
+const ns = __toModule(require_foo());
+let foo = 234;
+console.log(ns, ns.foo, foo);
 `,
 		},
 	})
@@ -617,85 +550,59 @@ func TestImportStarCommonJSNoCapture(t *testing.T) {
 			AbsOutputFile: "/out.js",
 		},
 		expected: map[string]string{
-			"/out.js": `bootstrap({
-  1(exports) {
-    // /foo.js
-    exports.foo = 123;
-  },
+			"/out.js": `// /foo.js
+var require_foo = __commonJS((exports) => {
+  exports.foo = 123;
+});
 
-  0() {
-    // /entry.js
-    const ns = __import(1 /* ./foo */);
-    let foo = 234;
-    console.log(ns.foo, ns.foo, foo);
-  }
-}, 0);
+// /entry.js
+const ns = __toModule(require_foo());
+let foo = 234;
+console.log(ns.foo, ns.foo, foo);
 `,
 		},
 	})
 }
 
-func TestImportStarES6AndCommonJS(t *testing.T) {
+func TestImportStarAndCommonJS(t *testing.T) {
 	expectBundled(t, bundled{
 		files: map[string]string{
-			"/entry1.js": `
+			"/entry.js": `
 				import * as ns from './foo'
-				console.log(ns.foo)
-			`,
-			"/entry2.js": `
-				const ns = require('./foo')
-				console.log(ns.foo)
+				const ns2 = require('./foo')
+				console.log(ns.foo, ns2.foo)
 			`,
 			"/foo.js": `
 				export const foo = 123
 			`,
 		},
-		entryPaths: []string{"/entry1.js", "/entry2.js"},
+		entryPaths: []string{"/entry.js"},
 		parseOptions: parser.ParseOptions{
 			IsBundling: true,
 		},
 		bundleOptions: BundleOptions{
-			IsBundling:   true,
-			AbsOutputDir: "/out",
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
 		},
 		expected: map[string]string{
-			"/out/entry1.js": `bootstrap({
-  2(exports) {
-    // /foo.js
-    __export(exports, {
-      foo: () => foo
-    });
-    const foo = 123;
-  },
+			"/out.js": `// /foo.js
+var require_foo = __commonJS((exports) => {
+  __export(exports, {
+    foo: () => foo2
+  });
+  const foo2 = 123;
+});
 
-  0() {
-    // /entry1.js
-    const ns = __import(2 /* ./foo */);
-    console.log(ns.foo);
-  }
-}, 0);
-`,
-			"/out/entry2.js": `bootstrap({
-  2(exports) {
-    // /foo.js
-    __export(exports, {
-      foo: () => foo
-    });
-    const foo = 123;
-  },
-
-  1() {
-    // /entry2.js
-    const ns = __require(2 /* ./foo */);
-    console.log(ns.foo);
-  }
-}, 1);
+// /entry.js
+const ns = __toModule(require_foo());
+const ns2 = require_foo();
+console.log(ns.foo, ns2.foo);
 `,
 		},
 	})
 }
 
-func TestImportStarES6NoBundleUnused(t *testing.T) {
+func TestImportStarNoBundleUnused(t *testing.T) {
 	expectBundled(t, bundled{
 		files: map[string]string{
 			"/entry.js": `
@@ -721,7 +628,7 @@ console.log(foo);
 	})
 }
 
-func TestImportStarES6NoBundleCapture(t *testing.T) {
+func TestImportStarNoBundleCapture(t *testing.T) {
 	expectBundled(t, bundled{
 		files: map[string]string{
 			"/entry.js": `
@@ -747,7 +654,7 @@ console.log(ns, ns.foo, foo);
 	})
 }
 
-func TestImportStarES6NoBundleNoCapture(t *testing.T) {
+func TestImportStarNoBundleNoCapture(t *testing.T) {
 	expectBundled(t, bundled{
 		files: map[string]string{
 			"/entry.js": `
@@ -773,7 +680,7 @@ console.log(ns.foo, ns.foo, foo);
 	})
 }
 
-func TestImportStarES6MangleNoBundleUnused(t *testing.T) {
+func TestImportStarMangleNoBundleUnused(t *testing.T) {
 	expectBundled(t, bundled{
 		files: map[string]string{
 			"/entry.js": `
@@ -800,7 +707,7 @@ console.log(foo);
 	})
 }
 
-func TestImportStarES6MangleNoBundleCapture(t *testing.T) {
+func TestImportStarMangleNoBundleCapture(t *testing.T) {
 	expectBundled(t, bundled{
 		files: map[string]string{
 			"/entry.js": `
@@ -827,7 +734,7 @@ console.log(ns, ns.foo, foo);
 	})
 }
 
-func TestImportStarES6MangleNoBundleNoCapture(t *testing.T) {
+func TestImportStarMangleNoBundleNoCapture(t *testing.T) {
 	expectBundled(t, bundled{
 		files: map[string]string{
 			"/entry.js": `
@@ -849,6 +756,1156 @@ func TestImportStarES6MangleNoBundleNoCapture(t *testing.T) {
 			"/out.js": `import * as ns from "./foo";
 let foo = 234;
 console.log(ns.foo, ns.foo, foo);
+`,
+		},
+	})
+}
+
+func TestImportStarExportStarOmitAmbiguous(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				import * as ns from './common'
+				console.log(ns)
+			`,
+			"/common.js": `
+				export * from './foo'
+				export * from './bar'
+			`,
+			"/foo.js": `
+				export const x = 1
+				export const y = 2
+			`,
+			"/bar.js": `
+				export const y = 3
+				export const z = 4
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /foo.js
+const x = 1;
+
+// /bar.js
+const z = 4;
+
+// /common.js
+const common_exports = {};
+__export(common_exports, {
+  x: () => x,
+  z: () => z
+});
+
+// /entry.js
+console.log(common_exports);
+`,
+		},
+	})
+}
+
+func TestImportExportStarAmbiguousError(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				import {x, y, z} from './common'
+				console.log(x, y, z)
+			`,
+			"/common.js": `
+				export * from './foo'
+				export * from './bar'
+			`,
+			"/foo.js": `
+				export const x = 1
+				export const y = 2
+			`,
+			"/bar.js": `
+				export const y = 3
+				export const z = 4
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expectedCompileLog: "/entry.js: error: Ambiguous import \"y\" has multiple matching exports\n",
+	})
+}
+
+func TestImportStarOfExportStarAs(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				import * as foo_ns from './foo'
+				console.log(foo_ns)
+			`,
+			"/foo.js": `
+				export * as bar_ns from './bar'
+			`,
+			"/bar.js": `
+				export const bar = 123
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /bar.js
+const bar_exports = {};
+__export(bar_exports, {
+  bar: () => bar
+});
+const bar = 123;
+
+// /foo.js
+const foo_exports = {};
+__export(foo_exports, {
+  bar_ns: () => bar_exports
+});
+
+// /entry.js
+console.log(foo_exports);
+`,
+		},
+	})
+}
+
+func TestImportOfExportStar(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				import {bar} from './foo'
+				console.log(bar)
+			`,
+			"/foo.js": `
+				export * from './bar'
+			`,
+			"/bar.js": `
+				// Add some statements to increase the part index (this reproduced a crash)
+				statement()
+				statement()
+				statement()
+				statement()
+				export const bar = 123
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /bar.js
+statement();
+statement();
+statement();
+statement();
+const bar = 123;
+
+// /foo.js
+
+// /entry.js
+console.log(bar);
+`,
+		},
+	})
+}
+
+func TestImportOfExportStarOfImport(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				import {bar} from './foo'
+				console.log(bar)
+			`,
+			"/foo.js": `
+				// Add some statements to increase the part index (this reproduced a crash)
+				statement()
+				statement()
+				statement()
+				statement()
+				export * from './bar'
+			`,
+			"/bar.js": `
+				export {value as bar} from './baz'
+			`,
+			"/baz.js": `
+				export const value = 123
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /baz.js
+const value = 123;
+
+// /bar.js
+
+// /foo.js
+statement();
+statement();
+statement();
+statement();
+
+// /entry.js
+console.log(value);
+`,
+		},
+	})
+}
+
+func TestExportSelfIIFE(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				export const foo = 123
+				export * from './entry'
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			OutputFormat:  printer.FormatIIFE,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `(() => {
+  // /entry.js
+  var require_entry = __commonJS((exports) => {
+    __export(exports, {
+      foo: () => foo
+    });
+    const foo = 123;
+  });
+  require_entry();
+})();
+`,
+		},
+	})
+}
+
+func TestExportSelfES6(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				export const foo = 123
+				export * from './entry'
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			OutputFormat:  printer.FormatESModule,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /entry.js
+const foo = 123;
+export {foo};
+`,
+		},
+	})
+}
+
+func TestExportSelfCommonJS(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				export const foo = 123
+				export * from './entry'
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			OutputFormat:  printer.FormatCommonJS,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /entry.js
+var require_entry = __commonJS((exports) => {
+  __export(exports, {
+    foo: () => foo
+  });
+  const foo = 123;
+});
+module.exports = require_entry();
+`,
+		},
+	})
+}
+
+func TestExportSelfCommonJSMinified(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				module.exports = {foo: 123}
+				console.log(require('./entry'))
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:        true,
+			MinifyIdentifiers: true,
+			OutputFormat:      printer.FormatCommonJS,
+			AbsOutputFile:     "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /entry.js
+var b = c((d, e) => {
+  e.exports = {
+    foo: 123
+  };
+  console.log(b());
+});
+module.exports = b();
+`,
+		},
+	})
+}
+
+func TestImportSelfCommonJS(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				exports.foo = 123
+				import {foo} from './entry'
+				console.log(foo)
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			OutputFormat:  printer.FormatCommonJS,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /entry.js
+var require_entry = __commonJS((exports) => {
+  const entry = __toModule(require_entry());
+  exports.foo = 123;
+  console.log(entry.foo);
+});
+module.exports = require_entry();
+`,
+		},
+	})
+}
+
+func TestExportSelfAsNamespaceES6(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				export const foo = 123
+				export * as ns from './entry'
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			OutputFormat:  printer.FormatESModule,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /entry.js
+const entry_exports = {};
+__export(entry_exports, {
+  foo: () => foo,
+  ns: () => entry_exports
+});
+const foo = 123;
+export {foo, entry_exports as ns};
+`,
+		},
+	})
+}
+
+func TestImportExportSelfAsNamespaceES6(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				export const foo = 123
+				import * as ns from './entry'
+				export {ns}
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			OutputFormat:  printer.FormatESModule,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /entry.js
+const entry_exports = {};
+__export(entry_exports, {
+  foo: () => foo,
+  ns: () => entry_exports
+});
+const foo = 123;
+export {foo, entry_exports as ns};
+`,
+		},
+	})
+}
+
+func TestReExportOtherFileExportSelfAsNamespaceES6(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				export * from './foo'
+			`,
+			"/foo.js": `
+				export const foo = 123
+				export * as ns from './foo'
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			OutputFormat:  printer.FormatESModule,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /foo.js
+const foo_exports = {};
+__export(foo_exports, {
+  foo: () => foo,
+  ns: () => foo_exports
+});
+const foo = 123;
+
+// /entry.js
+export {foo, foo_exports as ns};
+`,
+		},
+	})
+}
+
+func TestReExportOtherFileImportExportSelfAsNamespaceES6(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				export * from './foo'
+			`,
+			"/foo.js": `
+				export const foo = 123
+				import * as ns from './foo'
+				export {ns}
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			OutputFormat:  printer.FormatESModule,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /foo.js
+const foo_exports = {};
+__export(foo_exports, {
+  foo: () => foo,
+  ns: () => foo_exports
+});
+const foo = 123;
+
+// /entry.js
+export {foo, foo_exports as ns};
+`,
+		},
+	})
+}
+
+func TestOtherFileExportSelfAsNamespaceUnusedES6(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				export {foo} from './foo'
+			`,
+			"/foo.js": `
+				export const foo = 123
+				export * as ns from './foo'
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			OutputFormat:  printer.FormatESModule,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /foo.js
+const foo2 = 123;
+
+// /entry.js
+export {foo2 as foo};
+`,
+		},
+	})
+}
+
+func TestOtherFileImportExportSelfAsNamespaceUnusedES6(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				export {foo} from './foo'
+			`,
+			"/foo.js": `
+				export const foo = 123
+				import * as ns from './foo'
+				export {ns}
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			OutputFormat:  printer.FormatESModule,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /foo.js
+const foo2 = 123;
+
+// /entry.js
+export {foo2 as foo};
+`,
+		},
+	})
+}
+
+func TestExportSelfAsNamespaceCommonJS(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				export const foo = 123
+				export * as ns from './entry'
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			OutputFormat:  printer.FormatCommonJS,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /entry.js
+var require_entry = __commonJS((exports) => {
+  __export(exports, {
+    foo: () => foo,
+    ns: () => ns
+  });
+  const ns = __toModule(require_entry());
+  const foo = 123;
+});
+module.exports = require_entry();
+`,
+		},
+	})
+}
+
+func TestExportSelfAndRequireSelfCommonJS(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				export const foo = 123
+				console.log(require('./entry'))
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			OutputFormat:  printer.FormatCommonJS,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /entry.js
+var require_entry = __commonJS((exports) => {
+  __export(exports, {
+    foo: () => foo
+  });
+  const foo = 123;
+  console.log(require_entry());
+});
+module.exports = require_entry();
+`,
+		},
+	})
+}
+
+func TestExportSelfAndImportSelfCommonJS(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				import * as x from './entry'
+				export const foo = 123
+				console.log(x)
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			OutputFormat:  printer.FormatCommonJS,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /entry.js
+var require_entry = __commonJS((exports) => {
+  __export(exports, {
+    foo: () => foo
+  });
+  const x = __toModule(require_entry());
+  const foo = 123;
+  console.log(x);
+});
+module.exports = require_entry();
+`,
+		},
+	})
+}
+
+func TestExportOtherAsNamespaceCommonJS(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				export * as ns from './foo'
+			`,
+			"/foo.js": `
+				exports.foo = 123
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			OutputFormat:  printer.FormatCommonJS,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /foo.js
+var require_foo = __commonJS((exports2) => {
+  exports2.foo = 123;
+});
+
+// /entry.js
+__export(exports, {
+  ns: () => ns
+});
+const ns = __toModule(require_foo());
+`,
+		},
+	})
+}
+
+func TestImportExportOtherAsNamespaceCommonJS(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				import * as ns from './foo'
+				export {ns}
+			`,
+			"/foo.js": `
+				exports.foo = 123
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			OutputFormat:  printer.FormatCommonJS,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /foo.js
+var require_foo = __commonJS((exports2) => {
+  exports2.foo = 123;
+});
+
+// /entry.js
+__export(exports, {
+  ns: () => ns
+});
+const ns = __toModule(require_foo());
+`,
+		},
+	})
+}
+
+func TestNamespaceImportMissingES6(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				import * as ns from './foo'
+				console.log(ns, ns.foo)
+			`,
+			"/foo.js": `
+				export const x = 123
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /foo.js
+const foo_exports = {};
+__export(foo_exports, {
+  x: () => x
+});
+const x = 123;
+
+// /entry.js
+console.log(foo_exports, void 0);
+`,
+		},
+	})
+}
+
+func TestExportOtherCommonJS(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				export {bar} from './foo'
+			`,
+			"/foo.js": `
+				exports.foo = 123
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			OutputFormat:  printer.FormatCommonJS,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /foo.js
+var require_foo = __commonJS((exports2) => {
+  exports2.foo = 123;
+});
+
+// /entry.js
+__export(exports, {
+  bar: () => foo.bar
+});
+const foo = __toModule(require_foo());
+`,
+		},
+	})
+}
+
+func TestExportOtherNestedCommonJS(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				export {y} from './bar'
+			`,
+			"/bar.js": `
+				export {x as y} from './foo'
+			`,
+			"/foo.js": `
+				exports.foo = 123
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			OutputFormat:  printer.FormatCommonJS,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /foo.js
+var require_foo = __commonJS((exports2) => {
+  exports2.foo = 123;
+});
+
+// /bar.js
+const foo = __toModule(require_foo());
+
+// /entry.js
+__export(exports, {
+  y: () => foo.x
+});
+`,
+		},
+	})
+}
+
+func TestNamespaceImportUnusedMissingES6(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				import * as ns from './foo'
+				console.log(ns.foo)
+			`,
+			"/foo.js": `
+				export const x = 123
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /foo.js
+
+// /entry.js
+console.log(void 0);
+`,
+		},
+	})
+}
+
+func TestNamespaceImportMissingCommonJS(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				import * as ns from './foo'
+				console.log(ns, ns.foo)
+			`,
+			"/foo.js": `
+				exports.x = 123
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /foo.js
+var require_foo = __commonJS((exports) => {
+  exports.x = 123;
+});
+
+// /entry.js
+const ns = __toModule(require_foo());
+console.log(ns, ns.foo);
+`,
+		},
+	})
+}
+
+func TestNamespaceImportUnusedMissingCommonJS(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				import * as ns from './foo'
+				console.log(ns.foo)
+			`,
+			"/foo.js": `
+				exports.x = 123
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /foo.js
+var require_foo = __commonJS((exports) => {
+  exports.x = 123;
+});
+
+// /entry.js
+const ns = __toModule(require_foo());
+console.log(ns.foo);
+`,
+		},
+	})
+}
+
+func TestReExportNamespaceImportMissingES6(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				import {ns} from './foo'
+				console.log(ns, ns.foo)
+			`,
+			"/foo.js": `
+				export * as ns from './bar'
+			`,
+			"/bar.js": `
+				export const x = 123
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /bar.js
+const bar_exports = {};
+__export(bar_exports, {
+  x: () => x
+});
+const x = 123;
+
+// /foo.js
+
+// /entry.js
+console.log(bar_exports, bar_exports.foo);
+`,
+		},
+	})
+}
+
+func TestReExportNamespaceImportUnusedMissingES6(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				import {ns} from './foo'
+				console.log(ns.foo)
+			`,
+			"/foo.js": `
+				export * as ns from './bar'
+			`,
+			"/bar.js": `
+				export const x = 123
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /bar.js
+const bar_exports = {};
+__export(bar_exports, {
+  x: () => x
+});
+const x = 123;
+
+// /foo.js
+
+// /entry.js
+console.log(bar_exports.foo);
+`,
+		},
+	})
+}
+
+func TestNamespaceImportReExportMissingES6(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				import * as ns from './foo'
+				console.log(ns, ns.foo)
+			`,
+			"/foo.js": `
+				export {foo} from './bar'
+			`,
+			"/bar.js": `
+				export const x = 123
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expectedCompileLog: `/foo.js: error: No matching export for import "foo"
+/foo.js: error: No matching export for import "foo"
+`,
+	})
+}
+
+func TestNamespaceImportReExportUnusedMissingES6(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				import * as ns from './foo'
+				console.log(ns.foo)
+			`,
+			"/foo.js": `
+				export {foo} from './bar'
+			`,
+			"/bar.js": `
+				export const x = 123
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expectedCompileLog: `/foo.js: error: No matching export for import "foo"
+/foo.js: error: No matching export for import "foo"
+`,
+	})
+}
+
+func TestNamespaceImportReExportStarMissingES6(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				import * as ns from './foo'
+				console.log(ns, ns.foo)
+			`,
+			"/foo.js": `
+				export * from './bar'
+			`,
+			"/bar.js": `
+				export const x = 123
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /bar.js
+const x = 123;
+
+// /foo.js
+const foo_exports = {};
+__export(foo_exports, {
+  x: () => x
+});
+
+// /entry.js
+console.log(foo_exports, void 0);
+`,
+		},
+	})
+}
+
+func TestNamespaceImportReExportStarUnusedMissingES6(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				import * as ns from './foo'
+				console.log(ns.foo)
+			`,
+			"/foo.js": `
+				export * from './bar'
+			`,
+			"/bar.js": `
+				export const x = 123
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /bar.js
+
+// /foo.js
+
+// /entry.js
+console.log(void 0);
 `,
 		},
 	})
